requirements.txt
py-clob-client>=0.34.0
httpx>=0.27.0
python-dotenv>=1.0.0
web3>=6.0.0
